Merged PR 26203: merge ezpay-newrelic-deployment-marker-fix into master
temporary fix for ezpay api svc deployment markers (glq/grpc) needs to be tested in pipeline before merging.
possible TODO: create a list of "exception services" that match this case and loop through this logic for them.
i'll have to follow up with @<Steven Trunzo> on that when he's back in office.

Related work items: #221200, #221292